pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-aiops-triage"
version = "0.2.0"
description = "AIOps Triage Environment for OpenEnv — wraps Microsoft AIOpsLab and IBM ITBench benchmarks into a single uniform agent interface covering Detection, Localization, Analysis, and Mitigation tasks over simulated microservice clusters."
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.2",
    "fastapi>=0.104.0",
    "pydantic>=2.0",
    "uvicorn>=0.24.0",
    "requests>=2.25.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov", "httpx"]
inference = ["openai>=1.0.0"]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
packages = ["server"]
py-modules = ["models", "client"]
